services/parser/requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
olefile>=0.46
PyPDF2>=3.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
pymysql>=1.1.0
